new_backend/requirements.txt
google-play-scraper==1.2.4

# Core
fastapi
uvicorn[standard]
pydantic==2.6.4
python-dotenv==1.0.0


langchain-core==0.1.32
langchain-openai==0.0.8
langgraph>=0.0.30 # Ensure a recent version for checkpointer features

# LLM (Example: OpenAI)
openai>=1.0

# Database Checkpointing (Async PostgreSQL)
asyncpg>=0.28.0
psycopg # aiopg checkpointer might still need psycopg for parsing the DSN
SQLAlchemy>=2.0 # Often used indirectly or for utility with checkpointers
#psycopg2-binary aiopg
# Other potential DB drivers if needed elsewhere:
# psycopg2-binary

pydantic-ai

instructor>=0.4.8

pydantic_ai>=0.1.0

psycopg2-binary==2.9.9